requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
requests==2.32.3
google-genai>=1.30.0
